Lodgeit/requirements.txt
asgiref==3.8.1
defusedxml==0.7.1
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-xml==2.0.0
sqlparse==0.5.1
typing_extensions==4.12.2
whitenoise==6.8.1
